Changed to pass View to performAction
As suggested by rpdillon in his fork[1].

Note, this change breaks the api for performAction, if you have custom actions
they need to be migrated to the new interface. This is done by adding View as
an argument to performAction. The ToastAction in the example project now looks
like;

    private class ToastAction implements Action {

        @Override
        public int getDrawable() {
            return R.drawable.ic_title_export_default;
        }

        @Override
        public void performAction(View view) {
            Toast.makeText(OtherActivity.this,
                    "Example action", Toast.LENGTH_SHORT).show();
        }

    }

This also closes #5

[1] https://github.com/rpdillon/android-actionbar/commit/27da74326341b5ad7a3f1dbfd7ea091c36aae9be